rename diskstore -> filestore
