+++ storage/filestore/error.go
@@ -1,4 +1,4 @@
-package diskstore
+package filestore
 
 import "os"
